+++ xcube_smos/nckcindex/__init__.py
@@ -1,0 +1,1 @@
+from .nckcindex import NcKcIndex